Integrate mock server into server code
The server can now be run using the --mock command line option, which
causes it to use mock device controllers instead of the real ones.
Importing the device controllers is simpler now too. Mock hardware
settings are loaded in config.yml under the mock key for each device.
The underlying purpose behind doing this restructuring is because
loading the hardware state using a relative import is not compatible
with the setup as a Python package and this cleanly eliminates that
problem.
