+++ sctcamsoft/controllers/__init__.py
@@ -1,0 +1,3 @@
+from sctcamsoft.controllers.fan import *
+from sctcamsoft.controllers.network import *
+from sctcamsoft.controllers.power import *
