+++ sctcamsoft/controllers/mock/__init__.py
@@ -1,0 +1,3 @@
+from sctcamsoft.controllers.mock.fan import *
+from sctcamsoft.controllers.mock.network import *
+from sctcamsoft.controllers.mock.power import *
